requirements.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.0.0  # https://github.com/python-pillow/Pillow


# Django
# ------------------------------------------------------------------------------
django>=3    # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.11.0  # https://github.com/niwinz/django-redis
django-simple-history>=2.8
channels>=2.4  # https://channels.readthedocs.io/en/latest/installation.html
channels_redis >= 2.4 #https://pypi.org/project/channels-redis/

# Allows 'dangerous' queries
git+https://github.com/rob-deutsch/django-sql-explorer.git@django3 #https://github.com/groveco/django-sql-explorer

# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

# Test Add-ins
factory-boy==2.12.0  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==2.2  # https://github.com/jazzband/django-debug-toolbar
django-extensions==2.2.6  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==1.8.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==3.8.0  # https://github.com/pytest-dev/pytest-django